few more renamings
